mirror: Document (README.md) and use (Dockerfile) static builds
The README.md file explains that one can do a static build of bpftool by
passing CFLAGS=--static when invoking make, but that it does not work
out of the box when using libbfd. It should work, though; this is simply
a matter of passing the correct $(LIBS) flags, and we are missing
-liberty and -lz.

The reason why we are missing these is that the $(CFLAGS), containing
the --static flag, are not passed down to the feature detection system,
which does dynamic builds and finds that -liberty and -lz may not be
necessary.

However, the $(EXTRA_CFLAGS) variable is passed to the feature detection
system. If we use it instead of $(CFLAGS) to pass the --static flag,
then bpftool compiles just fine against libbfd.

Update the README.md file, and use static builds in the Dockerfile.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>